Move TDirectCopyActor to own source file
